Initial implementation for Windows
As noted in the documentation, this does not use the fancy Perl method
of URL encoding (yet; if it is changed to the documentation should be
updated again). So far I have only tested this on my Windows box so I
should pull it down on my Linux box and verify that it still works
there. Also it would be nice to get some verification from someone else
that it works on their Windows setup as well.

A little bit of cleanup may be necessary but I wanted to commit the
first working version.